feat(installer): deliver native Windows developer setup
Replace the generated PowerShell bootstrap with a statically linked native setup host and detached cleanup helper. Make the Windows package self-contained, transactional, maintainable, and suitable for clean developer machines while preserving Viper's zero-dependency product policy.

Native setup and lifecycle:
- add a branded high-DPI native wizard with current-user and all-users scope, validated custom destinations, component presets, disk preflight, accessible system theming, quiet/passive automation, and elevation only when required
- support install, upgrade, reinstall, modify, repair, explicit downgrade, and uninstall through one semantic-version-aware lifecycle engine with named-operation locks and Restart Manager process handling
- journal installation and removal, recover interrupted work, preserve ownership until commit, roll back registry, PATH, shortcuts, associations, and payload mutations, and use detached cleanup for residue-free self-removal
- persist selected components and integrations, publish developer prompt and CMake environment setup, safely open source files in ViperIDE, and expose maintenance, support, documentation, and update entries through Apps & Features

Packaging, trust, and reproducibility:
- introduce schema-3 canonical installer metadata for payload hashes, components, shortcuts, associations, source provenance, architecture, channel, maintenance identity, and authenticated update configuration
- build native host and cleanup targets with the static MSVC runtime while staging and verifying the architecture-matched app-local runtime required by installed tools and exported SDK libraries
- sign and verify nested Viper-owned payload executables before inventory hashing, preserve Microsoft-signed runtime files, embed signed maintenance helpers, recursively validate package trust, and sign the outer setup last
- separate stable and development identities, require clean immutable source metadata in release mode, honor SOURCE_DATE_EPOCH, reject stale or debug build artifacts, and emit deterministic manifests and artifact inventories
- add bounded same-origin HTTPS update discovery with pinned RSA-SHA256 signatures plus a canonical update-manifest authoring and verification script

Toolchain and integration hardening:
- carry UTF-8 filesystem paths through x64 and AArch64 code generation, archive discovery, cache lookup, native linking, and executable launch so Unicode Windows stages and install roots work end to end
- strengthen PE dependency, runtime import, compression, overlay, checksum, architecture, and native-host verification while keeping the package generator dependency-free
- let ViperIDE accept positional source files from safe Open With registrations and ship canonical installer artwork, manifests, version resources, shortcuts, file associations, samples, and developer documentation

Coverage and documentation:
- add focused metadata, version, update-signature, native-host launch, package CLI, archive-path, authoring-script, and packaging unit coverage
- expand user, elevated, XenoScape, production lifecycle, and fault-injection installer scenarios across clean install, upgrade, modify, repair, downgrade policy, cancellation, rollback, interrupted recovery, locked files, Unicode paths, native codegen, SDK discovery, and residue-free uninstall
- move 15 measured long-running Windows Debug tests to the opt-in slow lane, fix the update test's Common Controls activation context, and stabilize the serial Ridgebound D3D11 smoke ceiling without changing its release target
- record the installer-v2 architecture in ADR 0103 and update Windows onboarding, release, CLI, manpage, validation, update-signing, and test-lane documentation

Validation completed before this commit request: the Release Windows build/install pipeline passed 1,774 non-slow CTests; focused installer, packaging, update, user, XenoScape, production lifecycle, and destructive fault-injection coverage passed; and the generated x64 artifact passed strict installer validation. The final Windows Debug default lane passed 1,759/1,759 tests in 214.82 seconds, the migrated slow tests passed, and the update test passed 50 consecutive Debug runs plus final Debug and Release runs. Platform-policy lint and git diff checks were clean.

Tests were not rerun for this commit or the requested pull.

+++ src/tools/common/packaging/PkgDeflate.cpp
@@ -281,7 +281,7 @@ struct HuffmanTree {
             left -= blCount[bits];
             if (left < 0)
                 return false;
-            code = (code + blCount[bits - 1]) << 1;
+            code = static_cast<uint16_t>((code + blCount[bits - 1]) << 1);
             nextCode[bits] = code;
         }
 
@@ -559,7 +559,7 @@ static bool inflateDynamic(BitReader &br, OutputBuffer &out) {
 
     uint8_t clLengths[kMaxCodeLenCodes] = {};
     for (int i = 0; i < hclen; i++)
-        clLengths[kCodeLengthOrder[i]] = br.read(3);
+        clLengths[kCodeLengthOrder[i]] = static_cast<uint8_t>(br.read(3));
 
     HuffmanTree clTree;
     if (!clTree.build(clLengths, kMaxCodeLenCodes))
@@ -575,7 +575,7 @@ static bool inflateDynamic(BitReader &br, OutputBuffer &out) {
             return false;
 
         if (sym < 16) {
-            lengths[i++] = sym;
+            lengths[i++] = static_cast<uint8_t>(sym);
         } else if (sym == 16) {
             if (i == 0)
                 return false;
@@ -886,15 +886,15 @@ static void deflateFixed(BitWriter &bw, const uint8_t *data, size_t len, int lev
             int lenIdx = lenCode - 257;
 
             if (lenCode <= 279)
-                writeCode(bw, lenCode - 256, 7);
+                writeCode(bw, static_cast<uint16_t>(lenCode - 256), 7);
             else
-                writeCode(bw, 0xC0 + (lenCode - 280), 8);
+                writeCode(bw, static_cast<uint16_t>(0xC0 + (lenCode - 280)), 8);
 
             if (kLengthExtraBits[lenIdx] > 0)
                 bw.write(matchLen - kLengthBase[lenIdx], kLengthExtraBits[lenIdx]);
 
             int distCode = getDistCode(matchDist);
-            writeCode(bw, distCode, 5);
+            writeCode(bw, static_cast<uint16_t>(distCode), 5);
 
             if (kDistExtraBits[distCode] > 0)
                 bw.write(matchDist - kDistBase[distCode], kDistExtraBits[distCode]);
